Implement: eval config
Changes to be committed:
	modified:   eval_sim/robotwin/script/eval_policy_dp.py
	deleted:    scripts/mvdp/eval/dp_test_d345.sh
	deleted:    scripts/mvdp/eval/dp_test_d345_dp.sh
	new file:   scripts/mvdp/eval/dual_bottles_pick_hard_100_2v_23d_sav.sh
	new file:   scripts/mvdp/eval/dual_bottles_pick_hard_100_2v_3d_sav.sh
	new file:   scripts/mvdp/eval/dual_bottles_pick_hard_50_2v_23d.sh
	new file:   scripts/mvdp/eval/dual_bottles_pick_hard_50_2v_23d_sav.sh

+++ eval_sim/robotwin/script/eval_policy_dp.py
@@ -158,7 +158,7 @@ class DPRunner:
             action_dict = policy.predict_action(obs_dict_input)
 
         # device_transfer
-        np_action_dict = dict_apply(action_dict, lambda x: x.detach().to('cpu').numpy())
+        np_action_dict = dict_apply(action_dict, lambda x: x.detach().to('cpu').float().numpy())
         action = np_action_dict['action'].squeeze(0)
         return action
